# Conflicts:
#	src/main/java/de/srh/library/ui/borrowreturn/BorrowReturn.form
#	src/main/java/de/srh/library/ui/browselibrary/BrowseLibrary.form
#	src/main/java/de/srh/library/ui/createnewuser/CreateNewUser.form
#	src/main/java/de/srh/library/ui/editusers/EditUsers.form
#	src/main/java/de/srh/library/ui/enteremail/EnterEmail.form
#	src/main/java/de/srh/library/ui/login/LoginWindow.form
#	src/main/java/de/srh/library/ui/removebook/RemoveBook.form

+++ src/main/java/de/srh/library/ui/editbookdata/EditBookData.java
@@ -7,6 +7,7 @@ import de.srh.library.dto.ApiResponseCode;
 import de.srh.library.dto.BookDto;
 import de.srh.library.service.book.BookService;
 import de.srh.library.service.book.BookServiceImpl;
+import de.srh.library.ui.ConfirmationRequest;
 import de.srh.library.ui.login.LoginWindow;
 import de.srh.library.util.ValidatorUtils;
 import org.slf4j.Logger;
@@ -60,16 +61,29 @@ public class EditBookData extends JFrame {
         setSize(720, 1000);
         setDefaultCloseOperation(WindowConstants.DISPOSE_ON_CLOSE);
         setVisible(true);
+        toFront();
         logger.info("Opening edit book data window ...");
         loadCurrentBookData(bookId);
 
         saveChangesButton.addActionListener(new ActionListener() {
             @Override
             public void actionPerformed(ActionEvent e) {
+                try{
+                    inputValidation();
+                }catch (ValidateException ve ){
+                    JOptionPane.showMessageDialog(null, ve.getMessage());
+                    return;
+                }
 
-                JOptionPane.showMessageDialog(null, "Changes Saved!");
-                updateBookData(bookId);
-                saveChangesButton.setEnabled(false);
+                ConfirmationRequest confirmation = new ConfirmationRequest();
+                if(confirmation.userDecision) {
+                    updateBookData(bookId);
+                    saveChangesButton.setEnabled(false);
+                    JOptionPane.showMessageDialog(null, "Changes saved!");
+                }
+                else {
+                    JOptionPane.showMessageDialog(null, "Changes discarded!");
+                }
             }
         });
 
@@ -88,8 +102,15 @@ public class EditBookData extends JFrame {
         deleteBookButton.addActionListener(new ActionListener() {
             @Override
             public void actionPerformed(ActionEvent e) {
-                bookService = BookServiceImpl.createInstance();
-                bookService.removeBook(bookId);
+                ConfirmationRequest confirmation = new ConfirmationRequest();
+                if(confirmation.userDecision) {
+                    bookService = BookServiceImpl.createInstance();
+                    bookService.removeBook(bookId);
+                    JOptionPane.showMessageDialog(null, "Book deleted!");
+                }
+                else {
+                    JOptionPane.showMessageDialog(null, "Deletion canceled!");
+                }
             }
         });
     }
